+++ src/master.h
@@ -10,7 +10,6 @@
 #include <fcntl.h>
 #include <sys/resource.h>
 #include <sys/inotify.h>
-#include "fly.h"
 #include "fsignal.h"
 #include "conf.h"
 #include "alloc.h"
@@ -30,9 +29,12 @@ struct fly_master{
 	struct fly_pool_manager		*pool_manager;
 	struct fly_event_manager	*event_manager;
 
+	struct fly_bllist			signals;
+	size_t						sigcount;
 	struct fly_bllist			workers;
 	fly_context_t				*context;
 
+	struct fly_bllist			orig_sighandler;
 	struct fly_bllist			reload_filepath;
 	size_t						reload_pathcount;
 	fly_bit_t					detect_reload:1;
@@ -49,7 +51,6 @@ typedef struct fly_master fly_master_t;
 
 extern fly_signal_t fly_master_signals[];
 
-int fly_master_daemon(fly_context_t *ctx);
 int fly_create_pidfile(void);
 int fly_create_pidfile_noexit(void);
 void fly_remove_pidfile(void);
@@ -67,10 +68,6 @@ fly_context_t *fly_master_release_except_context(fly_master_t *master);
 __direct_log int fly_master_process(fly_master_t *master);
 void fly_master_worker_spawn(fly_master_t *master, void (*proc)(fly_context_t *, void *));
 
-#define __FLY_DEVNULL		("/dev/null")
-#define FLY_DAEMON_STDOUT	__FLY_DEVNULL
-#define FLY_DAEMON_STDERR	__FLY_DEVNULL
-#define FLY_DAEMON_STDIN	__FLY_DEVNULL
 #define fly_master_pid		getpid
 #define FLY_WORKER			"FLY_WORKER"
 #define FLY_WORKER_MAX		"FLY_WORKER_MAX"
@@ -79,5 +76,7 @@ void fly_master_worker_spawn(fly_master_t *master, void (*proc)(fly_context_t *,
 #define FLY_CREATE_PIDFILE					"FLY_CREATE_PIDFILE"
 bool fly_is_create_pidfile(void);
 void fly_master_setreload(fly_master_t *master, const char *reload_filepath, bool configure);
+
+#define FLY_NOTICE_WORKER_DAEMON_PID		(SIGRTMIN)
 
 #endif